teams separated by pools and ordered by ranking position
