+++ package.json
@@ -17,11 +17,13 @@
     "leaflet-defaulticon-compatibility": "^0.1.1",
     "leaflet-geosearch": "^3.7.0",
     "mdb-react-ui-kit": "^5.1.0",
+    "moment": "^2.29.4",
     "moment-timezone": "^0.5.41",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.2",
     "react-dom": "^18.2.0",
     "react-leaflet": "^4.2.1",
+    "react-moment": "^1.1.3",
     "react-router-bootstrap": "^0.26.2",
     "react-router-dom": "^6.9.0",
     "react-scripts": "5.0.1",
